# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Tesis Guvenlik & Operasyon API
  version: "0.1.0"
  description: |
    Multi-tenant tesis guvenlik & operasyon SaaS — v0 sozlesmesi.
    Backend (FastAPI) ile Mobil (Flutter) ve Panel (Next.js) gelistiricilerinin
    paralel calismasi icin tek dogruluk kaynagidir.

    Konvansiyonlar (detay: /contracts/README.md):
      * Tum zamanlar ISO8601 UTC (`...Z`). tenant.timezone ile yorumlanir.
      * Hata zarfi: `{ "error": { "code": "...", "message": "..." } }`.
      * Sayfalama: `limit` (varsayilan 50, max 200) + `offset`. Liste yanitlari
        `PageMeta` icerir.
      * Auth: JWT Bearer (bkz. /contracts/auth.md). Tenant her zaman token'dan
        cikarilir — istemci tenant_id gondermez/secemez.
      * `POST /scans` icin `Idempotency-Key` header zorunlu.
servers:
  # NOT: base path /v0 KALDIRILDI — gercek backend endpoint'leri /v0 ONEKI OLMADAN
  # sunuyor (orn. /auth/login, /scans). Base = host:port.
  - url: https://api.example.com
    description: production
  - url: http://localhost:8000
    description: yerel gelistirme
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: Kullanici listesi (admin)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/UserRole" }
        - name: is_active
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Ad/email icinde arama (opsiyonel)
          schema: { type: string }
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [users]
      summary: Kullanici olustur (admin)
      description: |
        Her rol (admin/yonetici/security/tesis_gorevlisi/resident) olusturulabilir. parola bcrypt ile
        hash'lenir; password_hash YANITTA donmez. email tenant icinde benzersiz.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: Olusturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: email bu tenant'ta zaten kayitli
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
        "422": { $ref: "#/components/responses/ValidationError" }
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [users]
      summary: Kullanici detay (admin)
      responses:
        "200":
          description: Detay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [users]
      summary: Kullanici guncelle (admin)
      description: |
        ad/telefon/role/is_active ve opsiyonel email/parola. parola verilirse yeniden
        hash'lenir. Kullanici SILINMEZ; pasiflestirme icin is_active=false kullanin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: Guncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: email bu tenant'ta zaten kayitli
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
        "422": { $ref: "#/components/responses/ValidationError" }

# =================================================================== #
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `Authorization: Bearer <token>`. Bkz. /contracts/auth.md"
  parameters:
    Limit:
      name: limit
      in: query
      description: Sayfa boyutu (varsayilan 50, max 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Atlanan kayit sayisi
      schema: { type: integer, minimum: 0, default: 0 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Gecersiz istek
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Kimlik dogrulama gerekli / gecersiz token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Rol bu islem icin yetkili degil
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kayit bulunamadi (tenant kapsaminda)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Govde dogrulama hatasi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Makine-okunabilir hata kodu
              example: validation_error
            message:
              type: string
              description: Insan-okunabilir aciklama
              example: "baslangic_saat bitis_saat'ten once olmali"
            details:
              type: array
              description: Alan bazli dogrulama hatalari (opsiyonel)
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    Page:
      type: object
      required: [meta]
      properties:
        meta: { $ref: "#/components/schemas/PageMeta" }
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
        total:
          type: integer
          description: Filtreye uyan toplam kayit sayisi
          example: 137
...
    UserRole:
      type: string
      enum: [admin, yonetici, security, tesis_gorevlisi, resident]

    # --------------------------- devices --------------------------- #
...
    User:
      type: object
      required: [id, ad, email, role, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        ad: { type: string }
        email: { type: string, format: email }
        telefon: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/UserRole" }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
    UserCreate:
      type: object
      required: [ad, email, role, password]
      properties:
        ad: { type: string }
        email: { type: string, format: email }
        telefon: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/UserRole" }
        password: { type: string, format: password, minLength: 8 }
    UserUpdate:
      type: object
      minProperties: 1
      properties:
        ad: { type: string }
        email: { type: string, format: email }
        telefon: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/UserRole" }
        is_active: { type: boolean }
        password: { type: string, format: password, minLength: 8 }
...